Added Google Authentication (#6)
* docs(README.md): add instructions to start the database and backend server using docker and maven respectively

* feat(authentication): add Google API call to validate access token and retrieve user information

* chore(pom.xml): add dependencies for jjwt, google-api-client, google-http-client, and google-oauth-client to support authentication and API calls

refactor(TokenValidator.java): update TokenValidator class to use GoogleIdTokenVerifier for token validation and extract user information from the token payload

* fix(TokenValidator.java): change JacksonFactory to GsonFactory to use Gson for JSON processing

The change was made to improve the JSON processing in the TokenValidator class. The JacksonFactory was replaced with the GsonFactory to use Gson library for JSON processing. This change ensures better compatibility and performance in handling JSON data.

* chore(pom.xml): remove unused dependencies to declutter the project and improve build performance